+++ api/main.py
@@ -4,7 +4,7 @@ from . import endpoints
 app = FastAPI(
     title="BrightWebApp API",
     description="A web API for the BrightWebApp package.",
-    version="1.0.0",
+    version="0.0.7",
 )
 
 app.include_router(endpoints.router)