+++ src/brightwebapp/brightway.py
@@ -1,4 +1,3 @@
-# %%
 from typing import Optional
 import bw2data as bd
 import bw2io as bi
@@ -7,6 +6,7 @@ import os
 def load_and_set_ecoinvent_project(
     username: Optional[str] = None,
     password: Optional[str] = None,
+    overwrite_existing: bool = False
 ) -> None:
     """Checks if the ecoinvent 3.10 Brightway project is installed.
     If not, loads it from Ecoinvent servers and installs it.
@@ -19,18 +19,27 @@ def load_and_set_ecoinvent_project(
     --------
     [`bw2io.bi.import_ecoinvent_release`](https://docs.brightway.dev/en/latest/content/api/bw2io/index.html#bw2io.import_ecoinvent_release)
     """
-    if "ei_3_10" not in bd.projects:
-        if username is None or password is None:
-            raise ValueError("Username and password must be provided to load ecoinvent project.")
+    project_name = 'ei_3_10'
+
+    # 1. if overwrite is requested, delete the project if it exists.
+    if overwrite_existing and project_name in bd.projects:
+        bd.projects.delete_project(project_name, delete_dir=True)
+
+    # 2. if the project doesn't exist, import it.
+    if project_name not in bd.projects:
+        if not username or not password:
+            raise ValueError("Username and password are required to download the ecoinvent database.")
         bi.import_ecoinvent_release(
             version='3.10',
             system_model='cutoff',
             username=username,
             password=password,
         )
-    else:
-        pass
-    bd.projects.set_current(name='ei_3_10')
+
+    # 3. now that the project is guaranteed to exist, set it as current.
+    bd.projects.set_current(project_name)
+
+    return 
 
 
 def load_and_set_useeio_project() -> None: